requirements.txt
allure_python_commons==2.13.5
pydantic==2.10.6
pytest==8.3.5
PyYAML==6.0.2
email-validator==2.2.0
python-dotenv==1.1.0
Requests==2.32.3
